#221 - rename external user DTO file & update refs
