+++ ccm_web/server/src/api/dtos/api.external.users.dto.ts
